Move `models.py` into `models` directory in `database_client`, rename to `core.py`
