+++ database_client/db_client_dataclasses.py
@@ -8,7 +8,7 @@ from sqlalchemy.schema import Column
 from sqlalchemy.sql.expression import asc, desc, BinaryExpression
 
 from database_client.enums import SortOrder
-from database_client.models import SQL_ALCHEMY_TABLE_REFERENCE
+from database_client.models.core import SQL_ALCHEMY_TABLE_REFERENCE
 
 
 ORDER_BY_REFERENCE = {
