+++ tests/integration/test_locations.py
@@ -5,7 +5,7 @@ from typing import Optional
 import pytest
 
 from database_client.enums import LocationType
-from database_client.models import Location
+from database_client.models.core import Location
 from middleware.schema_and_dto_logic.primary_resource_dtos.locations_dtos import (
     LocationPutDTO,
     LocationsGetRequestDTO,
