+++ tests/integration/test_proposals.py
@@ -1,10 +1,7 @@
 from http import HTTPStatus
-from http.client import HTTPException
-
-import pytest
 
 from database_client.enums import ApprovalStatus
-from database_client.models import Agency, LinkAgencyLocation
+from database_client.models.core import Agency, LinkAgencyLocation
 from middleware.enums import JurisdictionType, AgencyType
 from tests.helper_scripts.helper_classes.TestDataCreatorFlask import (
     TestDataCreatorFlask,
